Revert change to external zip header

+++ external/zip/src/zip.h
@@ -25,9 +25,9 @@ extern "C" {
 // 64-bit Windows is the only mainstream platform
 // where sizeof(long) != sizeof(void*)
 #ifdef _WIN64
-typedef signed __int64 ssize_t;
+typedef long long  ssize_t;  /* byte count or error */
 #else
-typedef signed int ssize_t;
+typedef long  ssize_t;  /* byte count or error */
 #endif
 #endif
